:construction: Adding restritions to director modification methods (requiring company code). :construction:

+++ Salestack/Controllers/Users/Director/DirectorsController.cs
@@ -20,17 +20,16 @@ public class DirectorsController : ControllerBase
     [HttpPost("companyCode={companyCode}")]
     public async Task<IActionResult> CreateDirectorAsync(SalestackDirector data, string companyCode)
     {
-        var companyForDirectorCreation = await _context.Company
-            .AsNoTracking()
+        var companyForDirectorCreateOperation = await _context.Company
             .FirstOrDefaultAsync(c => c.Id == data.CompanyId);
 
-        if (companyForDirectorCreation == null)
+        if (companyForDirectorCreateOperation == null)
             return BadRequest(new
             {
                 Message = $"No company found with id '{data.CompanyId}'."
             });
 
-        if (companyForDirectorCreation.CompanyCode != companyCode)
+        if (companyForDirectorCreateOperation.CompanyCode != companyCode)
             return BadRequest(new
             {
                 Message = "Please, provide a valid company code."
@@ -68,18 +67,34 @@ public class DirectorsController : ControllerBase
             .AsNoTracking()
             .FirstOrDefaultAsync(d => d.Id == id);
 
-        if (selectedDirector == null) return NotFound(new { Message = $"Director with id {id} not found." });
+        if (selectedDirector == null)
+            return NotFound(new
+            {
+                Message = $"Director with id {id} not found."
+            });
 
         return Ok(selectedDirector);
     }
 
-    [HttpPatch("{id}")]
-    public async Task<IActionResult> UpdateDirectorAsync(Guid id, SalestackDirector data)
+    [HttpPatch("{id}/companyCode={companyCode}")]
+    public async Task<IActionResult> UpdateDirectorAsync(Guid id, SalestackDirector data, string companyCode)
     {
         var selectedDirector = await _context.Director.FindAsync(id);
 
         if (selectedDirector == null)
-            return NotFound(new { Message = $"Director with id {id} not found." });
+            return NotFound(new
+            {
+                Message = $"Director with id {id} not found."
+            });
+
+        var companyForDirectorUpdateOperation = await _context.Company
+            .FirstOrDefaultAsync(c => c.Id == data.CompanyId);
+
+        if (companyForDirectorUpdateOperation!.CompanyCode != companyCode)
+            return BadRequest(new
+            {
+                Message = "Please, provide a valid company code."
+            });
 
         selectedDirector.Name = data.Name;
         selectedDirector.Email = data.Email;
@@ -90,15 +105,28 @@ public class DirectorsController : ControllerBase
         return Ok(selectedDirector);
     }
 
-    [HttpDelete("{id}")]
-    public async Task<IActionResult> DeleteDirectorAsync(Guid id)
+    [HttpDelete("{id}/companyCode={companyCode}")]
+    public async Task<IActionResult> DeleteDirectorAsync(Guid id, string companyCode)
     {
         var selectedDirector = await _context.Director.FindAsync(id);
 
         if (selectedDirector == null)
-            return NotFound(new { Message = $"Director with id {id} not found." });
+            return NotFound(new
+            {
+                Message = $"Director with id {id} not found."
+            });
+
+        var companyFromWichDirectorWillBeDeleted = await _context.Company
+           .FirstOrDefaultAsync(c => c.Id == selectedDirector.CompanyId);
+
+        if (companyFromWichDirectorWillBeDeleted!.CompanyCode != companyCode)
+            return BadRequest(new
+            {
+                Message = "Please, provide a valid company code."
+            });
 
         _context.Director.Remove(selectedDirector);
+
         await _context.SaveChangesAsync();
 
         return NoContent();
